tests: move branch tests to different file and add more
